Allocate the kernels array as a float64
The kernels array should be float64, otherwise when summing the
evaluations on the nodes for each cell generates numerical errors.

+++ SimPEG/potential_fields/gravity/simulation.py
@@ -252,7 +252,7 @@ def fill_sensitivity_matrix(
     for i in prange(n_receivers):
         receiver_index = receivers_indices[i]
         # Allocate vector for kernels evaluated on mesh nodes
-        kernels = np.empty(n_nodes, dtype=dtype)
+        kernels = np.empty(n_nodes)
         for j in range(n_nodes):
             x, y, z = receivers[receiver_index, :] - nodes[j, :]
             distance = np.sqrt(x**2 + y**2 + z**2)
